requirements.txt
bottle==0.12.19
jsonschema==4.4.0
psycopg2-binary==2.9.3
